requirements.txt
asgiref==3.7.2
Django==3.0.3
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==19.1.0
numpy==1.25.1
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
